+++ src/button_input.h
@@ -6,6 +6,8 @@ class DisplayEngine;
 
 typedef void (*PresetChangeCallback)(int newIndex, void *context);
 
+bool buttonInputRecoveryHeldAtBoot();
+
 void buttonInputBegin(DisplayEngine *engine, PresetChangeCallback callback = nullptr,
                       void *callbackContext = nullptr);
 void buttonInputTick();