+++ src/wifi_manager.h
@@ -2,7 +2,7 @@
 
 #include <Arduino.h>
 
-bool wifiManagerBegin();
+bool wifiManagerBegin(bool forceRecoveryAp = false);
 void wifiManagerForgetSta();
 bool wifiManagerConnectSta(const char *ssid, const char *password);
 bool wifiManagerHasSavedSta();
@@ -14,3 +14,7 @@ String wifiStaSsid();
 String wifiStaIp();
 int wifiStaRssi();
 void wifiManagerTick();
+bool wifiManagerApActive();
+bool wifiManagerRecoveryMode();
+int wifiManagerApClientCount();
+const char *wifiManagerModeString();